+++ libfreeipmi/include/freeipmi/interface/ipmi-lan-interface.h
@@ -40,16 +40,16 @@ extern fiid_template_t tmpl_lan_msg_hdr_rq;
 extern fiid_template_t tmpl_lan_msg_hdr_rs;
 extern fiid_template_t tmpl_lan_msg_trlr;
 
-int8_t fill_lan_session_hdr (uint8_t authentication_type,
-                             uint32_t session_sequence_number,
-                             uint32_t session_id,
-                             fiid_obj_t obj_lan_session_hdr);
+int fill_lan_session_hdr (uint8_t authentication_type,
+                          uint32_t session_sequence_number,
+                          uint32_t session_id,
+                          fiid_obj_t obj_lan_session_hdr);
 
-int8_t fill_lan_msg_hdr (uint8_t rs_addr,
-                         uint8_t net_fn,
-                         uint8_t rs_lun,
-                         uint8_t rq_seq,
-                         fiid_obj_t obj_lan_msg_hdr);
+int fill_lan_msg_hdr (uint8_t rs_addr,
+                      uint8_t net_fn,
+                      uint8_t rs_lun,
+                      uint8_t rq_seq,
+                      fiid_obj_t obj_lan_msg_hdr);
 
 int32_t assemble_ipmi_lan_pkt (fiid_obj_t obj_rmcp_hdr,
                                fiid_obj_t obj_lan_session_hdr,
@@ -60,13 +60,13 @@ int32_t assemble_ipmi_lan_pkt (fiid_obj_t obj_rmcp_hdr,
                                uint8_t *pkt,
                                uint32_t pkt_len);
 
-int8_t unassemble_ipmi_lan_pkt (uint8_t *pkt,
-                                uint32_t pkt_len,
-                                fiid_obj_t obj_rmcp_hdr,
-                                fiid_obj_t obj_lan_session_hdr,
-                                fiid_obj_t obj_lan_msg_hdr,
-                                fiid_obj_t obj_cmd,
-                                fiid_obj_t obj_lan_msg_trlr);
+int unassemble_ipmi_lan_pkt (uint8_t *pkt,
+                             uint32_t pkt_len,
+                             fiid_obj_t obj_rmcp_hdr,
+                             fiid_obj_t obj_lan_session_hdr,
+                             fiid_obj_t obj_lan_msg_hdr,
+                             fiid_obj_t obj_cmd,
+                             fiid_obj_t obj_lan_msg_trlr);
 
 ssize_t ipmi_lan_sendto (int s,
                          const void *buf,
