+++ libfreeipmi/include/freeipmi/interface/ipmi-rmcpplus-interface.h
@@ -191,51 +191,51 @@ extern fiid_template_t tmpl_rmcpplus_rakp_message_4;
  *
  * Returns 0 on success, -1 on error.
  */
-int8_t ipmi_rmcpplus_init (void);
+int ipmi_rmcpplus_init (void);
 
-int8_t fill_rmcpplus_session_hdr (uint8_t payload_type,
-                                  uint8_t payload_authenticated,
-                                  uint8_t payload_encrypted,
-                                  uint32_t oem_iana,
-                                  uint16_t oem_payload_id,
-                                  uint32_t session_id,
-                                  uint32_t session_sequence_number,
-                                  fiid_obj_t obj_rmcpplus_session_hdr);
+int fill_rmcpplus_session_hdr (uint8_t payload_type,
+                               uint8_t payload_authenticated,
+                               uint8_t payload_encrypted,
+                               uint32_t oem_iana,
+                               uint16_t oem_payload_id,
+                               uint32_t session_id,
+                               uint32_t session_sequence_number,
+                               fiid_obj_t obj_rmcpplus_session_hdr);
 
-int8_t fill_rmcpplus_session_trlr (fiid_obj_t obj_rmcpplus_session_trlr);
+int fill_rmcpplus_session_trlr (fiid_obj_t obj_rmcpplus_session_trlr);
 
-int8_t fill_rmcpplus_payload (uint8_t *confidentiality_header,
-                              uint32_t confidentiality_header_len,
-                              uint8_t *payload_data,
-                              uint32_t payload_data_len,
-                              uint8_t *confidentiality_trailer,
-                              uint32_t confidentiality_trailer_len,
-                              fiid_obj_t obj_cmd_rq);
+int fill_rmcpplus_payload (uint8_t *confidentiality_header,
+                           uint32_t confidentiality_header_len,
+                           uint8_t *payload_data,
+                           uint32_t payload_data_len,
+                           uint8_t *confidentiality_trailer,
+                           uint32_t confidentiality_trailer_len,
+                           fiid_obj_t obj_cmd_rq);
 
-int8_t fill_rmcpplus_open_session (uint8_t message_tag,
-                                   uint8_t requested_maximum_privilege_level,
-                                   uint32_t remote_console_session_id,
-                                   uint8_t authentication_algorithm,
-                                   uint8_t integrity_algorithm,
-                                   uint8_t confidentiality_algorithm,
-                                   fiid_obj_t obj_cmd_rq);
+int fill_rmcpplus_open_session (uint8_t message_tag,
+                                uint8_t requested_maximum_privilege_level,
+                                uint32_t remote_console_session_id,
+                                uint8_t authentication_algorithm,
+                                uint8_t integrity_algorithm,
+                                uint8_t confidentiality_algorithm,
+                                fiid_obj_t obj_cmd_rq);
 
-int8_t fill_rmcpplus_rakp_message_1 (uint8_t message_tag,
-                                     uint32_t managed_system_session_id,
-                                     uint8_t *remote_console_random_number,
-                                     uint32_t remote_console_random_number_len,
-                                     uint8_t requested_maximum_privilege_level,
-                                     uint8_t name_only_lookup_flag,
-                                     char *username,
-                                     uint32_t username_len,
-                                     fiid_obj_t obj_cmd_rq);
+int fill_rmcpplus_rakp_message_1 (uint8_t message_tag,
+                                  uint32_t managed_system_session_id,
+                                  uint8_t *remote_console_random_number,
+                                  uint32_t remote_console_random_number_len,
+                                  uint8_t requested_maximum_privilege_level,
+                                  uint8_t name_only_lookup_flag,
+                                  char *username,
+                                  uint32_t username_len,
+                                  fiid_obj_t obj_cmd_rq);
 
-int8_t fill_rmcpplus_rakp_message_3 (uint8_t message_tag,
-                                     uint8_t rmcpplus_status_code,
-                                     uint32_t managed_system_session_id,
-                                     uint8_t *key_exchange_authentication_code,
-                                     uint32_t key_exchange_authentication_code_len,
-                                     fiid_obj_t obj_cmd_rq);
+int fill_rmcpplus_rakp_message_3 (uint8_t message_tag,
+                                  uint8_t rmcpplus_status_code,
+                                  uint32_t managed_system_session_id,
+                                  uint8_t *key_exchange_authentication_code,
+                                  uint32_t key_exchange_authentication_code_len,
+                                  fiid_obj_t obj_cmd_rq);
 
 int32_t assemble_ipmi_rmcpplus_pkt (uint8_t authentication_algorithm,
                                     uint8_t integrity_algorithm,
@@ -254,22 +254,22 @@ int32_t assemble_ipmi_rmcpplus_pkt (uint8_t authentication_algorithm,
                                     uint8_t *pkt,
                                     uint32_t pkt_len);
 
-int8_t unassemble_ipmi_rmcpplus_pkt (uint8_t authentication_algorithm,
-                                     uint8_t integrity_algorithm,
-                                     uint8_t confidentiality_algorithm,
-                                     uint8_t *integrity_key,
-                                     uint32_t integrity_key_len,
-                                     uint8_t *confidentiality_key,
-                                     uint32_t confidentiality_key_len,
-                                     uint8_t *pkt,
-                                     uint32_t pkt_len,
-                                     fiid_obj_t obj_rmcp_hdr,
-                                     fiid_obj_t obj_rmcpplus_session_hdr,
-                                     fiid_obj_t obj_rmcpplus_payload,
-                                     fiid_obj_t obj_lan_msg_hdr,
-                                     fiid_obj_t obj_cmd,
-                                     fiid_obj_t obj_lan_msg_trlr,
-                                     fiid_obj_t obj_rmcpplus_session_trlr);
+int unassemble_ipmi_rmcpplus_pkt (uint8_t authentication_algorithm,
+                                  uint8_t integrity_algorithm,
+                                  uint8_t confidentiality_algorithm,
+                                  uint8_t *integrity_key,
+                                  uint32_t integrity_key_len,
+                                  uint8_t *confidentiality_key,
+                                  uint32_t confidentiality_key_len,
+                                  uint8_t *pkt,
+                                  uint32_t pkt_len,
+                                  fiid_obj_t obj_rmcp_hdr,
+                                  fiid_obj_t obj_rmcpplus_session_hdr,
+                                  fiid_obj_t obj_rmcpplus_payload,
+                                  fiid_obj_t obj_lan_msg_hdr,
+                                  fiid_obj_t obj_cmd,
+                                  fiid_obj_t obj_lan_msg_trlr,
+                                  fiid_obj_t obj_rmcpplus_session_trlr);
 
 #ifdef __cplusplus
 }
